+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- `Uuid::parse_me`, to parse mixed-endian UUID strings.
+
+### Changed
+
+- Documented how `Uuid::from_bytes_me` works better
+- Documented `Uuid` `Debug` representation
+- `Uuid::version` and `Uuid::variant` are now `const fn`
+
 ## [0.4.0] - 2022-08-02
 
 ### Added