+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.4.1] - 2022-08-03
+
 ### Added
 
 - `Uuid::parse_me`, to parse mixed-endian UUID strings.
@@ -89,7 +91,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - no_std UUID's
 
 <!-- next-url -->
-[Unreleased]: https://github.com/DianaNites/nuuid/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/DianaNites/nuuid/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/DianaNites/nuuid/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/DianaNites/nuuid/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/DianaNites/nuuid/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/DianaNites/nuuid/compare/v0.3.0...v0.3.1